Extracted common functions from all different landingzone and moved to the utils file
