+++ Tombolo/server/jobs/lzMonitoring/lzFileCountMonitoring.js
@@ -4,7 +4,10 @@ const { decryptString } = require('../../utils/cipher');
 const { FileSprayService } = require('@hpcc-js/comms');
 const { getClusterOptions } = require('../../utils/getClusterOptions');
 const { generateNotificationId } = require('../jobMonitoring/monitorJobsUtil');
-const { getFilesFromSingleLzDirectory } = require('./lzFileMonitoringUtils');
+const {
+  getFilesFromSingleLzDirectory,
+  findLocalDateTimeAtCluster,
+} = require('./lzFileMonitoringUtils');
 
 const {
   landingZoneMonitoring: LandingZoneMonitoring,
@@ -14,13 +17,6 @@ const {
   asr_products: AsrProducts,
   asr_domains: AsrDomains,
 } = models;
-
-// Local time at cluster
-function findLocalDateTimeAtCluster(timeZoneOffset) {
-  const newDate = new Date();
-  const localTime = new Date(newDate.getTime() + timeZoneOffset * 60 * 1000);
-  return localTime;
-}
 
 const monitoring_name = 'Landing Zone Monitoring';
